Refactoring TabWidget and TabButtonWidget to a different directory. #54
