+++ lib/src/internal/tabs_area/tabs_area.dart
@@ -7,7 +7,7 @@ import '../../tabbed_view_controller.dart';
 import '../../theme/tabbed_view_theme_data.dart';
 import '../../theme/tabs_area_theme_data.dart';
 import '../../theme/theme_widget.dart';
-import '../tab_widget.dart';
+import '../tab/tab_widget.dart';
 import '../tabbed_view_provider.dart';
 import 'hidden_tabs.dart';
 import 'tabs_area_corner.dart';
